requirements.txt
python==3.9
streamlit==0.90.0
PyMuPDF==1.18.11
python-docx==0.8.11
